# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inea Scents API",
        "description": "API Documentation for Inea Scents",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get user's wishlist",
                "operationId": "25e51433040f79ef596ee4001198d4f2",
                "responses": {
                    "200": {
                        "description": "List of wishlist packages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Package"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/wishlist/toggle": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Toggle package in wishlist",
                "operationId": "c690ac875ef9156df71fdce102c96867",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "package_id"
                                ],
                                "properties": {
                                    "package_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Toggle status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "attached": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Package": {
                "title": "Package",
                "description": "Package model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Romantic Getaway"
                    },
                    "description": {
                        "type": "string",
                        "example": "A nice package for couples."
                    },
                    "inclusions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "pax_options": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "freebies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "reviews_count": {
                        "type": "integer",
                        "example": 120
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "gallery_images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "scents": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Scent"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Scent": {
                "title": "Scent",
                "description": "Scent model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lavender"
                    },
                    "description": {
                        "type": "string",
                        "example": "A calming floral scent."
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/lavender.jpg"
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}